Restructured RecordingPage Folder
Changed component structure within  the Recording Page folder. Fixed not-matching css stylesheet names.
