+++ frontend/src/RecordingPageComp/InstrumentSelection/InstrumentSelection.css
