+++ frontend/src/RecordingPageComp/InstrumentSelection/InstrumentSelectionComp.jsx
