+++ frontend/src/RecordingPageComp/RecordingComp/RecordingComp.css
