+++ frontend/src/RecordingPageComp/RecordingPage/RecordingPage.css
